refactor(tui): split committed-band into commit + repin modules
committed-band.ts (599) -> thin barrel re-exporting two cohesive modules:
- committed-band-repin.ts (133): flushResizeGhostErase + repositionCommittedBand
  (post-render / post-resize band re-pinning) -- <350.
- committed-band-commit.ts (474): the commitAbove commit pipeline + clear/reset.
  Stays >350 because commitAbove is a single irreducible ~389-line pipeline;
  splitting it would thread 14 shared locals through a parameter object -- a
  risky logic change on the gap-regression-prone commit path, not a net win.
Behavior-preserving (functions moved verbatim; only imports + barrel changed).
